backend/requirements.txt
Flask==0.10.1
#Application server for both development and production
gunicorn==19.4.5
#Wrapper for Twitter's API
tweepy==3.8.0
#.Env file reader
python-dotenv==0.14.0
# CORS for FrontEnd calls
flask-cors==3.0.8
werkzeug==0.16.1

# Testing
pytest==2.9.1
pytest-cov==2.2.1
# Linter
flake8==2.5.4
#CLI
Click==6.4
